Switch from MaterialDialogs to AppCompat dialogs

+++ app/src/main/java/com/tortel/authenticator/dialog/ConfirmDeleteDialog.java
@@ -1,14 +1,15 @@
 package com.tortel.authenticator.dialog;
 
 import android.app.Dialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
 import android.support.v4.content.LocalBroadcastManager;
+import android.support.v7.app.AlertDialog;
 import android.text.Html;
 import android.widget.TextView;
 
-import com.afollestad.materialdialogs.MaterialDialog;
 import com.tortel.authenticator.R;
 import com.tortel.authenticator.activity.MainActivity;
 import com.tortel.authenticator.common.data.AccountDb;
@@ -43,35 +44,31 @@ public class ConfirmDeleteDialog extends DialogFragment {
 
     @Override
     public Dialog onCreateDialog(Bundle savedInstanceState) {
-        MaterialDialog.Builder builder = new MaterialDialog.Builder(getActivity());
-        builder.title(getString(R.string.remove_account_dialog_title, username));
+        AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+        builder.setTitle(getString(R.string.remove_account_dialog_title, username));
 
         TextView textView = new TextView(getActivity());
         textView.setText(Html.fromHtml(getString(R.string.remove_account_dialog_message)));
 
-        builder.customView(textView, true);
+        builder.setView(textView);
 
-        builder.positiveText(R.string.ok);
-        builder.negativeText(R.string.cancel);
-        builder.callback(callback);
+        builder.setPositiveButton(R.string.ok, callback);
+        builder.setNegativeButton(R.string.cancel, callback);
 
-        return builder.build();
+        return builder.create();
     }
 
-    private MaterialDialog.ButtonCallback callback = new MaterialDialog.ButtonCallback() {
+    private DialogInterface.OnClickListener callback = new DialogInterface.OnClickListener() {
         @Override
-        public void onNegative(MaterialDialog materialDialog) {
-            dismiss();
-        }
-
-        @Override
-        public void onPositive(MaterialDialog materialDialog) {
-            AccountDb accountDb = DependencyInjector.getAccountDb();
-            accountDb.delete(id);
-            Intent intent = new Intent(MainActivity.ACCOUNT_DELETED);
-            intent.putExtra(MainActivity.ACCOUNT_ID, id);
-            LocalBroadcastManager broadcastManager = LocalBroadcastManager.getInstance(getActivity().getBaseContext());
-            broadcastManager.sendBroadcast(intent);
+        public void onClick(DialogInterface dialog, int which) {
+            if (which == DialogInterface.BUTTON_POSITIVE) {
+                AccountDb accountDb = DependencyInjector.getAccountDb();
+                accountDb.delete(id);
+                Intent intent = new Intent(MainActivity.ACCOUNT_DELETED);
+                intent.putExtra(MainActivity.ACCOUNT_ID, id);
+                LocalBroadcastManager broadcastManager = LocalBroadcastManager.getInstance(getActivity());
+                broadcastManager.sendBroadcast(intent);
+            }
 
             dismiss();
         }
